pythonProject/requirements.txt
gql==3.3.0
Jinja2==3.1.2
Pillow==9.1.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.0
sendgrid==6.9.7
xhtml2pdf==0.2.8
